test: Add test for automatic direct node routing
[ upstream commit 150cf7bf05a12f458330fe77031e8e49c8a0eeb0 ]

This restructures the tunnel tests into a more generic datapath configuration
test suite and adds a dedicated test for automatic direct node routes.
