+++ test/k8sT/manifests/cilium-ds-patch-auto-node-routes.yaml
@@ -10,7 +10,8 @@ spec:
         name: cilium-agent
         args:
         - "--debug=$(CILIUM_DEBUG)"
-        - "--tunnel=geneve"
+        - "--tunnel=disabled"
+        - "--auto-direct-node-routes"
         - "--kvstore=etcd"
         - "--kvstore-opt=etcd.config=/var/lib/etcd-config/etcd.config"
         - "--k8s-require-ipv4-pod-cidr"
