+++ test/k8sT/manifests/cilium-ds-patch-vxlan.yaml
@@ -10,10 +10,11 @@ spec:
         name: cilium-agent
         args:
         - "--debug=$(CILIUM_DEBUG)"
-        - "--tunnel=disabled"
-        - "--auto-routing"
+        - "--tunnel=vxlan"
         - "--kvstore=etcd"
         - "--kvstore-opt=etcd.config=/var/lib/etcd-config/etcd.config"
+        - "--k8s-require-ipv4-pod-cidr"
+        - "--pprof=true"
         - "--log-system-load"
       volumes:
       - name: etcd-secrets
